+++ widgets/src/lib.rs
@@ -18,13 +18,12 @@ pub mod scrollbar;
 pub mod select_region;
 pub mod slider;
 pub mod tabs;
-pub mod text_field;
 
 pub mod transform_box;
 pub mod prelude {
   pub use super::{
     avatar::*, buttons::*, checkbox::*, common_widget::*, divider::*, grid_view::*, icon::*,
     input::*, label::*, layout::*, list::*, menu::*, path::*, progress::*, radio::*, router::*,
-    scrollbar::*, select_region::*, slider::*, tabs::*, text_field::*, transform_box::*,
+    scrollbar::*, select_region::*, slider::*, tabs::*, transform_box::*,
   };
 }